Setup

+++ database/migrations/2016_07_11_073130_create_geolocations_table.php
@@ -14,8 +14,8 @@ class CreateGeolocationsTable extends Migration
     {
         Schema::create('geolocations', function (Blueprint $table) {
             $table->increments('id');
-            $table->double('longitude');
-            $table->double('latitude');
+            $table->float('longitude');
+            $table->float('latitude');
         });
     }
